+++ tests/test_model.py
@@ -307,3 +307,7 @@ class TestModel(TestQueryBase):
     def test_multiple_primary_keys(self):
         self.assertEqual(MultiplePrimaryKey.primary_key, inspect(MultiplePrimaryKey).primary_key)
 
+    def test_chaining(self):
+        foo = Foo.get(1)
+        self.assertIs(foo.update({}).flush().save().expire().refresh().expunge(), foo)
+