+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 - Add `ModelBase.primary_key` class property for retrieving primary key(s).
 - Add `Base=None` argument to `make_declarative_base()` to support passing in a subclass of `ModelBase`. Previously had to create a declarative `Model` to pass in a subclassed `ModelBase`.
+- Add chaining support for `ModelBase` methods: `update`, `flush`, `save`, `expire`, `refresh`, and `expunge`.
 
 ## v0.4.2 (2014-02-24)